+++ test/plugin/scenarios/httpexchange-scenario/httpexchange-projectA-scenario/src/main/java/org/apache/skywalking/apm/testcase/sc/httpexchange/projectA/api/TestcaseClient.java
@@ -1,0 +1,43 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+
+package org.apache.skywalking.apm.testcase.sc.httpexchange.projectA.api;
+
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.service.annotation.GetExchange;
+import org.springframework.web.service.annotation.HttpExchange;
+import org.springframework.web.service.annotation.PostExchange;
+
+@HttpExchange("/testcase")
+public interface TestcaseClient {
+
+    @GetExchange("/healthCheck")
+    String healthCheck();
+
+    @GetExchange("/success")
+    String success();
+
+    @GetExchange("/error")
+    String error();
+
+    @GetExchange("/url/{test}")
+    String urlParams(@PathVariable("test") String test);
+
+    @PostExchange(value = "/body")
+    String bodyParams(@RequestBody String body);
+}