+++ test/plugin/scenarios/httpexchange-scenario/httpexchange-projectA-scenario/src/main/java/org/apache/skywalking/apm/testcase/sc/httpexchange/projectA/config/RestClientConfig.java
@@ -1,0 +1,43 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+
+package org.apache.skywalking.apm.testcase.sc.httpexchange.projectA.config;
+
+import org.apache.skywalking.apm.testcase.sc.httpexchange.projectA.api.TestcaseClient;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.reactive.function.client.WebClient;
+import org.springframework.web.reactive.function.client.support.WebClientAdapter;
+import org.springframework.web.service.invoker.HttpServiceProxyFactory;
+
+@Configuration
+public class RestClientConfig {
+    @Value("${projectB.host:localhost:18080}")
+    private String hostbAddress;
+
+    @Bean
+    public TestcaseClient testcaseClient() {
+        WebClient client = WebClient.builder()
+                                    .baseUrl("http://" + hostbAddress + "/")
+                                    .build();
+        HttpServiceProxyFactory factory = HttpServiceProxyFactory
+            .builderFor(WebClientAdapter.create(client))
+            .build();
+        return factory.createClient(TestcaseClient.class);
+    }
+}